% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMatSolve.R
\name{fMatSolve}
\alias{fMatSolve}
\title{Replaced base::solve Function}
\usage{
fMatSolve(a, b, tol = .Machine$double.eps, LINPACK = FALSE, ...)
}
\arguments{
\item{a, b, tol, LINPACK}{Please refer to \link[base]{solve}.}
}
\description{
Replaced base::solve Function
}
