% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMKLVersion}
\alias{getMKLVersion}
\title{Function to get the version of Intel MKL}
\usage{
getMKLVersion()
}
\value{
The version of Intel MKL
}
\description{
Function to get the version of Intel MKL
}
\examples{
getMKLVersion()
}
