% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMKL.MatrixCal-package.R
\docType{package}
\name{onemkl-package}
\alias{onemkl-package}
\title{`oneMKL.MatrixCal` Package}
\description{
The `oneMKL.MatrixCal` package provide matrix operation functions based on Intel® oneAPI Math Kernel Library (oneMKL).
}
