% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mkl_threads}
\alias{mkl_threads}
\alias{setMKLThreads}
\alias{getMKLThreads}
\title{Function to get/set the number of threads used in Intel MKL}
\usage{
setMKLThreads(nThreads)

getMKLThreads()
}
\arguments{
\item{nThreads}{The number of threads you want to use in Intel MKL.}
}
\value{
The number of threads.
}
\description{
Function to get/set the number of threads used in Intel MKL
}
\examples{
getMKLThreads() # Default is the number of CPUs cores on your PC

\dontrun{
setMKLThreads(1)
getMKLThreads() # 1
}
}
